Feature/view counter#khama9

+++ Assets/HoloMeApp/Scripts/UI/UIBtnLikes.cs
@@ -7,13 +7,18 @@ using TMPro;
 
 namespace Beem.UI {
 
+
+    /// <summary>
+    /// Need for demonstration count of likes
+    /// can invoke like and unlike event
+    /// </summary>
     public class UIBtnLikes : MonoBehaviour {
         [SerializeField] Image imageLike;
         [SerializeField] Image imageUnlike;
         [SerializeField] TMP_Text likesCount;
 
         private bool _isLike;
-        private int _count;
+        private long _count;
         private long _streamId = -1;
 
         /// <summary>
@@ -47,7 +52,7 @@ namespace Beem.UI {
             }
         }
 
-        private void UpdateState(long streamId, bool isLike, int count) {
+        private void UpdateState(long streamId, bool isLike, long count) {
             if (streamId != _streamId)
                 return;
             _isLike = isLike;
